fix(auth): merge
